+++ node/src/GlideClient.ts
@@ -14,6 +14,7 @@ import {
     createConfigSet,
     createCustomCommand,
     createEcho,
+    createFunctionLoad,
     createInfo,
     createPing,
     createSelect,
@@ -302,12 +303,41 @@ export class GlideClient extends BaseClient {
      *
      * @example
      * ```typescript
-     * // Example usage of time method without any argument
+     * // Example usage of time command
      * const result = await client.time();
      * console.log(result); // Output: ['1710925775', '913580']
      * ```
      */
     public time(): Promise<[string, string]> {
         return this.createWritePromise(createTime());
+    }
+
+    /**
+     * Loads a library to Valkey.
+     *
+     * See https://valkey.io/commands/function-load/ for details.
+     *
+     * since Valkey version 7.0.0.
+     *
+     * @param libraryCode - The source code that implements the library.
+     * @param replace - Whether the given library should overwrite a library with the same name if it
+     *     already exists.
+     * @returns The library name that was loaded.
+     *
+     * @example
+     * ```typescript
+     * // Example usage of `FUNCTION LOAD` command
+     * const code = "#!lua name=mylib \n redis.register_function('myfunc', function(keys, args) return args[1] end)";
+     * const result = await client.functionLoad(code, true);
+     * console.log(result); // Output: 'mylib'
+     * ```
+     */
+    public functionLoad(
+        libraryCode: string,
+        replace?: boolean,
+    ): Promise<string> {
+        return this.createWritePromise(
+            createFunctionLoad(libraryCode, replace),
+        );
     }
 }