+++ node/tests/RedisClient.test.ts
@@ -15,11 +15,12 @@ import { v4 as uuidv4 } from "uuid";
 import { GlideClient, ProtocolVersion, Transaction } from "..";
 import { RedisCluster } from "../../utils/TestUtils.js";
 import { command_request } from "../src/ProtobufMessage";
-import { runBaseTests } from "./SharedTests";
+import { checkIfServerVersionLessThan, runBaseTests } from "./SharedTests";
 import {
     checkSimple,
     convertStringArrayToBuffer,
     flushAndCloseClient,
+    generateLuaLibCode,
     getClientConfigurationOption,
     intoString,
     parseCommandLineArgs,
@@ -294,6 +295,110 @@ describe("GlideClient", () => {
             }
 
             client.close();
+        },
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        "function load test_%p",
+        async (protocol) => {
+            if (await checkIfServerVersionLessThan("7.0.0")) return;
+
+            const client = await GlideClient.createClient(
+                getClientConfigurationOption(cluster.getAddresses(), protocol),
+            );
+
+            try {
+                const libName = "mylib1C" + uuidv4().replace(/-/g, ""); //.replaceAll("-", "");
+                const funcName = "myfunc1c" + uuidv4().replace(/-/g, ""); //.replaceAll("-", "");
+                const code = generateLuaLibCode(
+                    libName,
+                    new Map([[funcName, "return args[1]"]]),
+                    true,
+                );
+                // TODO use commands instead of customCommand once implemented
+                // verify function does not yet exist
+                expect(
+                    await client.customCommand([
+                        "FUNCTION",
+                        "LIST",
+                        "LIBRARYNAME",
+                        libName,
+                    ]),
+                ).toEqual([]);
+
+                expect(await client.functionLoad(code)).toEqual(
+                    Buffer.from(libName),
+                );
+
+                expect(
+                    await client.customCommand([
+                        "FCALL",
+                        funcName,
+                        "0",
+                        "one",
+                        "two",
+                    ]),
+                ).toEqual(Buffer.from("one"));
+                expect(
+                    await client.customCommand([
+                        "FCALL_RO",
+                        funcName,
+                        "0",
+                        "one",
+                        "two",
+                    ]),
+                ).toEqual(Buffer.from("one"));
+
+                // TODO verify with FUNCTION LIST
+                // re-load library without replace
+
+                await expect(client.functionLoad(code)).rejects.toThrow(
+                    `Library '${libName}' already exists`,
+                );
+
+                // re-load library with replace
+                expect(await client.functionLoad(code, true)).toEqual(
+                    Buffer.from(libName),
+                );
+
+                // overwrite lib with new code
+                const func2Name = "myfunc2c" + uuidv4().replaceAll("-", "");
+                const newCode = generateLuaLibCode(
+                    libName,
+                    new Map([
+                        [funcName, "return args[1]"],
+                        [func2Name, "return #args"],
+                    ]),
+                    true,
+                );
+                expect(await client.functionLoad(newCode, true)).toEqual(
+                    Buffer.from(libName),
+                );
+
+                expect(
+                    await client.customCommand([
+                        "FCALL",
+                        func2Name,
+                        "0",
+                        "one",
+                        "two",
+                    ]),
+                ).toEqual(2);
+                expect(
+                    await client.customCommand([
+                        "FCALL_RO",
+                        func2Name,
+                        "0",
+                        "one",
+                        "two",
+                    ]),
+                ).toEqual(2);
+            } finally {
+                expect(
+                    await client.customCommand(["FUNCTION", "FLUSH"]),
+                ).toEqual("OK");
+                client.close();
+            }
         },
     );